Зарефакторил сервисы после ревью, накидал todo для завершения getDirList
